Fixing a small bug in window exit behaviour in getResIntEnGUI.py
Former-commit-id: b37e52dc41490ef508d152b56831771da15dbe75

+++ getresinten/getResIntEnGUI.py
@@ -169,12 +169,11 @@ class DesignInteractCalculate(QtWidgets.QMainWindow,design.Ui_MainWindow):
 
 	def stopCalculation(self):
 		# Parse the log file for any child PID spawned by getResIntEn.py
-		print('detected stop')
 		if hasattr(self,"processGetResIntEn"):
 
-			print('supposed to kill the running process.')
-			# Stop the calculation
-			os.kill(self.processGetResIntEn.pid,signal.SIGINT)
+			if self.processGetResIntEn:
+				# Stop the calculation
+				os.kill(self.processGetResIntEn.pid,signal.SIGINT)
 
 			# Stop monitoring
 			self.stopMonitorProgressThread.emit()
